// src/enums/equipmentId.ts
export enum EquipmentId {
  SHORT_SWORD = "shortSword",
  LONG_SWORD = "longSword",
  SMALL_SHIELD = "smallShield",
  LARGE_SHIELD = "largeShield",
  LIGHT_MAIL = "lightMail",
  LIGHT_GLOVES = "lightGloves",
  LIGHT_BOOTS = "lightBoots",
  LIGHT_HELMET = "lightHelmet",
  HEAVY_MAIL = "heavyMail",
  HEAVY_GLOVES = "heavyGloves",
  HEAVY_BOOTS = "heavyBoots",
  HEAVY_HELMET = "heavyHelmet",
}